requirements.txt
docker-compose==1.5.2
docker-py==1.5.0
dockerpty==0.3.4
docopt==0.6.1
enum34==1.0.4
functools32==3.2.3.post2
jsonschema==2.5.1
pygraphviz==1.3.1
PyYAML==3.11
requests==2.7.0
six==1.7.3
texttable==0.8.4
websocket-client==0.32.0
wheel==0.24.0
